+++ packages/Dropdown/src/Dropdown.stories.tsx
@@ -94,16 +94,33 @@ export const DropdownInAModal = () => {
   return (
     <>
       <Button onClick={() => setShow(true)}>open</Button>
-      <Modal isClosable isOpen={show} onClose={() => setShow(false)}>
-        <div style={{ height: 400, overflow: 'auto' }}>
+      <Modal isClosable isDismissable isOpen={show} onClose={() => setShow(false)}>
+        <div style={{ height: 400 }}>
           Scroll down
           <div style={{ height: 1000, paddingTop: 500 }}>
             <Dropdown
               isOpen={open}
               onClose={() => setOpen(false)}
+              isScrollable
               content={
                 <List
                   items={[
+                    'Organization details',
+                    'Billing',
+                    'Permissions',
+                    'Segments',
+                    'Organization details',
+                    'Billing',
+                    'Permissions',
+                    'Segments',
+                    'Organization details',
+                    'Billing',
+                    'Permissions',
+                    'Segments',
+                    'Organization details',
+                    'Billing',
+                    'Permissions',
+                    'Segments',
                     'Organization details',
                     'Billing',
                     'Permissions',