+++ packages/TagPicker/src/TagPicker.tsx
@@ -299,6 +299,7 @@ const TagPicker: React.FunctionComponent<TagPickerProps> = ({
         dismissible={!disabled}
         appearance="secondary"
         onRemove={handleTagRemove}
+        hasError={s.hasError}
       >
         {s.text}
       </Tag>
@@ -404,6 +405,7 @@ const TagPicker: React.FunctionComponent<TagPickerProps> = ({
       isOpen={shouldShowResults}
       onClose={handleLoseFocus}
       onClick={handleGainFocus}
+      isScrollable
     >
       {tagPickerElem}
     </Dropdown>